Refactor based on PR (#258)
