+++ apps/api/src/view-models/buildViewModelForResource/viewModels/coscrad-user.view-model.ts
@@ -1,4 +1,4 @@
-import { IUserViewModel } from '@coscrad/api-interfaces';
+import { ICoscradUserViewModel } from '@coscrad/api-interfaces';
 import { CoscradUserRole, FromDomainModel } from '@coscrad/data-types';
 import { CoscradUserProfile } from '../../../domain/models/user-management/user/entities/user/coscrad-user-profile.entity';
 import { CoscradUser } from '../../../domain/models/user-management/user/entities/user/coscrad-user.entity';
@@ -6,7 +6,7 @@ import { BaseViewModel } from './base.view-model';
 
 const FromUser = FromDomainModel(CoscradUser);
 
-export class CoscradUserViewModel extends BaseViewModel implements IUserViewModel {
+export class CoscradUserViewModel extends BaseViewModel implements ICoscradUserViewModel {
     @FromUser
     readonly profile: CoscradUserProfile;
