+++ apps/coscrad-frontend/src/store/slices/auth/constants.ts
@@ -1,1 +1,1 @@
-export const AUTH = 'authorization';
+export const AUTH = 'auth';
