+++ apps/coscrad-frontend/src/store/slices/auth/selectors/select-authenticated-user.ts
@@ -2,10 +2,10 @@ import { RootState } from '../../..';
 import { AUTH } from '../constants';
 
 export const selectAuthenticatedUser = (state: RootState) => {
-    const { userInfo, hasAuthenticatedUser } = state[AUTH];
+    const { userAuthInfo, hasAuthenticatedUser } = state[AUTH];
 
     if (!hasAuthenticatedUser) return null;
 
     // TODO join in actual user profile
-    return userInfo;
+    return userAuthInfo;
 };
