+++ apps/coscrad-frontend/src/store/slices/auth/types/auth-slice-state.ts
@@ -2,5 +2,5 @@ import { AuthenticatedUserInfo } from './authenticated-user-info';
 
 export type AuthSliceState = {
     hasAuthenticatedUser: boolean;
-    userInfo: AuthenticatedUserInfo;
+    userAuthInfo: AuthenticatedUserInfo;
 };
