+++ libs/api-interfaces/src/lib/aggregate-views/view-models/resources/coscrad-user/coscrad-user.view-model.interface.ts
@@ -1,0 +1,7 @@
+import { CoscradUserRole } from './coscrad-user-role';
+
+export interface ICoscradUserViewModel {
+    username: string;
+    // TODO Make this CoscradUserRole
+    roles: CoscradUserRole[];
+}
