+++ libs/api-interfaces/src/lib/aggregate-views/view-models/resources/coscrad-user/index.ts
@@ -1,0 +1,2 @@
+export * from './coscrad-user-role';
+export * from './coscrad-user.view-model.interface';
