+++ libs/api-interfaces/src/lib/aggregate-views/view-models/resources/index.ts
@@ -1,6 +1,7 @@
 export * from '../categorizable-type.enum';
 export * from './bibliographic-reference';
 export * from './book';
+export * from './coscrad-user';
 export * from './media-items';
 export * from './photograph.view-model.interface';
 export * from './resource-composite-identifier';
@@ -10,5 +11,4 @@ export * from './spatial-feature';
 export * from './term.view-model.interface';
 export * from './transcribed-audio.view-model.interface';
 export * from './types';
-export * from './user.view-model.interface';
 export * from './vocabulary-list';
