--- libs/api-interfaces/src/lib/aggregate-views/view-models/resources/user.view-model.interface.ts
@@ -1,5 +1,0 @@
-export interface IUserViewModel {
-    username: string;
-    // TODO Make this CoscradUserRole
-    roles: string[];
-}
